Refactor/tests (#7172)
* improving some tests

* refactors

* refactors
