+++ conans/test/functional/command/info/info_folders_test.py
